requirements.txt
llama-cpp-python
xmltodict
requests
